+++ src/libponyrt/asio/event.h
@@ -14,7 +14,7 @@ PONY_EXTERN_C_BEGIN
  */
 typedef struct asio_event_t
 {
-	intptr_t fd;          /* file descriptor */
+	int fd;          /* file descriptor */
 	uint32_t eflags;      /* event filter flags */
 	pony_actor_t* owner;  /* owning actor */
 	uint32_t msg_id;      /* I/O handler (actor message) */
@@ -39,7 +39,7 @@ typedef struct asio_msg_t
  *  An event is noisy, if it should prevent the runtime system from terminating
  *  based on quiescence.
  */
-asio_event_t* asio_event_create(intptr_t fd, uint32_t eflags, uint32_t msg_id,
+asio_event_t* asio_event_create(int fd, uint32_t eflags, uint32_t msg_id,
 	bool noisy, void* udata);
 
 /** Deallocates an ASIO event.
